Taking out error disabling setting `set +e`. The tagging didn't work as expected. Also was not able to push artifact to github release. Received a permissions error. Regenerated token and updated taggin script as well. Removed quotes around encrypted token.
Troubling shooting tagging script on CI. Using quotes around encrypted
token to see if this resolves permissions problem.

Tested the if statement with a command and corrected the if statement
logic.

Finally decided to changed api key to secret env because token is not
working. And updated logging to tagging script.